requirements.txt
google-api-python-client
google-cloud-storage==2.7.0
requests